Extracted handleDeepResearchStartFailure to failure-handler.ts
Moved the function and its types to a standalone module with no Supabase
static imports. The test now imports directly from failure-handler.ts,
avoiding the Bun 1.3.14 module corruption that occurred when start.ts's
full import chain (including Supabase clients) was loaded in CI without env vars.
